+++ backend/controllers/profiles.js
@@ -2,8 +2,9 @@
 const router = require('express').Router()
 
 // ROUTES
+// GET list of profiles
 router.get('/', (req, res) => {
-    res.send('GET /users')
+    res.render('profiles/index')
 })
 
 // export