+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## [0.0.1] 2022-10-22
+### Added
+* basic profiles index
+* basic default, error404, and home views
+
+### Changed
+* "users" changed to "profiles" because "users" will be used for the password and account stuff
+
 ## [0.0.0] 2022-10-20
 ### Added
 * CHANGELOG.md